move logic to server side
